Typeorm의 entity와 migration file 추가 및 수정

+++ packages/typeorm/src/migration/1573489925976-tags.ts
@@ -1,30 +1,51 @@
-import { MigrationInterface, QueryRunner, Table } from "typeorm";
+import { MigrationInterface, QueryRunner, Table } from 'typeorm';
+import { Status } from '../custom-type';
 
 export class tags1573489925976 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<any> {
-    await queryRunner.createTable(
-      new Table({
-        name: "tags",
-        columns: [
-          {
-            name: "id",
-            isPrimary: true,
-            type: "int",
-            isGenerated: true,
-            generationStrategy: "increment"
-          },
-          {
-            name: "name",
-            type: "varchar",
-            length: "45"
-          }
-        ]
-      }),
-      true
-    );
+    if (!(await queryRunner.hasTable('tags'))) {
+      await queryRunner.createTable(
+        new Table({
+          name: 'tags',
+          columns: [
+            {
+              name: 'id',
+              isPrimary: true,
+              type: 'int',
+              isGenerated: true,
+              generationStrategy: 'increment',
+            },
+            {
+              name: 'name',
+              type: 'varchar',
+              length: '45',
+            },
+            {
+              name: 'createdAt',
+              type: 'datetime',
+              default: 'CURRENT_TIMESTAMP',
+            },
+            {
+              name: 'updateAt',
+              type: 'datetime',
+              default: 'CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP',
+            },
+            {
+              name: 'status',
+              type: 'tinyint',
+              length: '4',
+              default: Status.active,
+            },
+          ],
+        }),
+        true,
+      );
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<any> {
-    await queryRunner.dropTable("tags", true, true);
+    if (await queryRunner.hasTable('tags')) {
+      await queryRunner.dropTable('tags', true, true);
+    }
   }
 }
